+++ CHANGELOG.md
@@ -39,6 +39,8 @@ Recorder.FileType changed from method to write-only property
 - Recorder.Prefix
 - Recorder.Eject() references 'Command.Eject'
 - Recorder.State
+- Command.Reset()
+- Command.Save($filepath)
 
 ### Changed